chore(menu): change export to menu.ts
