+++ src/demo-app/menu/menu-demo.ts
@@ -1,5 +1,5 @@
 import {Component} from '@angular/core';
-import {MD_MENU_DIRECTIVES} from '@angular2-material/menu/menu-trigger';
+import {MD_MENU_DIRECTIVES} from '@angular2-material/menu/menu';
 import {MD_ICON_DIRECTIVES} from '@angular2-material/icon/icon';
 import {MD_BUTTON_DIRECTIVES} from '@angular2-material/button/button';
 import {MD_TOOLBAR_DIRECTIVES} from '@angular2-material/toolbar/toolbar';
